requirements.txt
python_telegram_bot==12.0.0b1
gTTS==2.0.3
PyYAML==5.1
telegram==0.0.1
requests==2.21.0